# swagger/swagger.yaml
definitions:
  dto.AgentInboundRequest:
    properties:
      jid:
        description: Jid is the WhatsApp chat identifier this message came from.
        type: string
      text:
        description: Text is the raw inbound message text.
        type: string
    required:
    - jid
    - text
    type: object
  dto.AgentInboundResponse:
    properties:
      jobId:
        description: |-
          JobID is the correlation ID the bot should register as a pending
          job — the eventual /callback/extract-done call will carry this
          same ID.
        type: string
      status:
        allOf:
        - $ref: '#/definitions/dto.ExtractStatus'
        description: Status indicates the current state of the agent turn.
    type: object
  dto.ApifyResource:
    properties:
      defaultDatasetId:
        description: DefaultDatasetId is the ID of the dataset created by the Actor
          run.
        type: string
      status:
        description: Status indicates the final status of the Actor run (e.g., SUCCEEDED,
          FAILED).
        type: string
    required:
    - defaultDatasetId
    - status
    type: object
  dto.ApifyWebhookPayload:
    properties:
      eventType:
        description: EventType indicates the type of event that triggered the webhook.
        type: string
      resource:
        allOf:
        - $ref: '#/definitions/dto.ApifyResource'
        description: Resource contains details of the Apify Actor run that just finished.
    required:
    - eventType
    - resource
    type: object
...
  dto.ExtractAPIRequest:
    properties:
      question:
        description: Question is a free-text question to search the knowledge base.
        type: string
      url:
        description: URL is the target media link provided by the bot.
        type: string
    type: object
  dto.ExtractData:
    properties:
      transcript:
        description: Transcript contains the transcribed text from the tiktok video.
        type: string
    type: object
  dto.ExtractResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/dto.ExtractData'
        description: Data contains the final extraction results. It is omitted if
          the job is not yet completed.
      error:
        description: Error contains the error message if the extraction failed. It
          is omitted otherwise.
        type: string
      jobId:
        description: JobID is the unique identifier for the extraction task.
        type: string
      status:
        allOf:
        - $ref: '#/definitions/dto.ExtractStatus'
        description: Status indicates the current state of the extraction job.
    type: object
  dto.ExtractStatus:
    enum:
    - processing
    - completed
    - failed
    type: string
    x-enum-varnames:
    - StatusProcessing
    - StatusCompleted
    - StatusFailed
...
  model.APIResponse:
    properties:
      data: {}
      error:
        type: string
      request_id:
        format: uuid
        type: string
      status:
        maximum: 599
        minimum: 100
        type: integer
    required:
    - request_id
    - status
    type: object
...
info:
  contact: {}
paths:
  /api/agent/inbound:
    post:
      consumes:
      - application/json
      description: |-
        Single entry point for every inbound WhatsApp message. Applies a
        deterministic fast path for bare TikTok URLs (straight to the existing
        cache-check/Apify pipeline) and routes everything else through the
        conversational agent (tool-calling over claim-checking and video
        processing, with per-jid conversation history).
      parameters:
      - description: Inbound Message
        in: body
        name: form
        required: true
        schema:
          $ref: '#/definitions/dto.AgentInboundRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            allOf:
            - $ref: '#/definitions/model.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.AgentInboundResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/model.APIResponse'
            - properties:
                error:
                  type: string
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/model.APIResponse'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/model.APIResponse'
            - properties:
                error:
                  type: string
              type: object
      summary: Handle an inbound WhatsApp message
      tags:
      - agent
  /api/extract:
    post:
      consumes:
      - application/json
      description: Enqueues a job to extract media from a target URL using Apify.
      parameters:
      - description: Extraction Request Form
        in: body
        name: form
        required: true
        schema:
          $ref: '#/definitions/dto.ExtractAPIRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            allOf:
            - $ref: '#/definitions/model.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ExtractResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/model.APIResponse'
            - properties:
                error:
                  type: string
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/model.APIResponse'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/model.APIResponse'
            - properties:
                error:
                  type: string
              type: object
      summary: Trigger media extraction
      tags:
      - media
...
  /api/webhooks/apify:
    post:
      consumes:
      - application/json
      description: Receives successful run notifications from Apify.
      parameters:
      - description: Apify Webhook Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.ApifyWebhookPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.APIResponse'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/model.APIResponse'
            - properties:
                error:
                  type: string
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/model.APIResponse'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/model.APIResponse'
            - properties:
                error:
                  type: string
              type: object
      summary: Apify webhook receiver
      tags:
      - webhooks
swagger: "2.0"